+++ src/master/JobTracker.java
@@ -8,13 +8,13 @@ import java.util.HashMap;
  *
  */
 public class JobTracker {
-	private static HashMap<Integer, Job> jobMap = new HashMap<Integer, Job>();  //Store map of jobs
+	private static HashMap<Long, Job> jobMap = new HashMap<Long, Job>();  //Store map of jobs
 	
-	public static void addJob(int jobID, Job job) {
+	public static void addJob(long jobID, Job job) {
 		jobMap.put(jobID, job);
 	}
 	
-	public static Job getJob(int jobID) {
+	public static Job getJob(long jobID) {
 		return jobMap.get(jobID);
 	}
