+++ src/slave/AssignmentHandler.java
@@ -19,7 +19,7 @@ public interface AssignmentHandler extends Remote {
 	 * @param jobAssigner   Get the master's job assigner. In order to upload data to master
 	 * @throws RemoteException
 	 */
-	public int addAssignment(int nodeID, int jobID, ArrayList<Integer> repList, JobAssigner jobAssigner) throws RemoteException;
-	public byte[] uploadResult(int jobID, int repNum) throws RemoteException;
+	public int addAssignment(int nodeID, long jobID, ArrayList<Integer> repList, JobAssigner jobAssigner) throws RemoteException;
+	public byte[] uploadResult(long jobID, int repNum) throws RemoteException;
 	public void addRep(ArrayList<Integer> repList) throws RemoteException;
 }
